fix: refactorisation de l'api en modules
