+++ apps/api/src/app.middleware.ts
@@ -1,7 +1,7 @@
 import { JwtService } from '@nestjs/jwt';
 import { Injectable, NestMiddleware, HttpException, HttpStatus } from '@nestjs/common';
 import { Request, Response, NextFunction } from 'express';
-import { UserService } from './service/user.service';
+import { UserService } from './user/user.service';
 
 interface UserRequest extends Request {
 	user: any
