+++ apps/api/src/prediction/prediction.schema.ts
