+++ apps/api/src/publication/publication.schema.ts
