+++ apps/api/src/vote/vote.schema.ts
